+++ packages/bundler/src/db/statusdb.ts
@@ -1,6 +1,6 @@
 import IORedis from "ioredis";
 import { RedisTransaction } from ".";
-import { OperationStatus } from "../common";
+import { OperationStatus } from "@nocturne-xyz/sdk";
 
 const JOB_STATUS_PREFIX = "JOB_STATUS_";