+++ packages/bundler/src/router.ts
@@ -1,13 +1,12 @@
 import { Queue } from "bullmq";
 import IORedis from "ioredis";
 import {
-  OperationStatus,
   ProvenOperationJobData,
   PROVEN_OPERATION_QUEUE,
   PROVEN_OPERATION_JOB_TAG,
 } from "./common";
 import { Request, Response } from "express";
-import { calculateOperationDigest, ProvenOperation } from "@nocturne-xyz/sdk";
+import { OperationStatus, calculateOperationDigest, ProvenOperation } from "@nocturne-xyz/sdk";
 import { OperationValidator } from "./validator";
 import * as JSON from "bigint-json-serialization";
 import { StatusDB } from "./db";