+++ packages/sdk/src/commonTypes.ts
@@ -244,3 +244,12 @@ export interface PreProofOperation extends BaseOperation {
 export interface ProvenOperation extends BaseOperation {
   joinSplitTxs: ProvenJoinSplitTx[];
 }
+
+export enum OperationStatus {
+  QUEUED = "QUEUED",
+  PRE_BATCH = "PRE_BATCH",
+  IN_BATCH = "IN_BATCH",
+  IN_FLIGHT = "IN_FLIGHT",
+  EXECUTED_SUCCESS = "EXECUTED_SUCCESS",
+  EXECUTED_FAILED = "EXECUTED_FAILED",
+}